Desktop PreviewStrut, etc. refactored for online reuse
PreviewStrut, ZoneVectorBall, and StrutBuilderController are now all safe for
use via JSweet, in preparation for implementing the strut builder tool for
classic online.

Desktop strut building works, and unit tests pass.
